+++ CHANGELOG.md
@@ -4,6 +4,28 @@ All notable changes to this project are documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.2.0] - 2023-04-xx
+
+### Added
+
+* Version info in the panel footer
+* Quality check
+  * dimensions: are they equal and power of 2?
+  * resolution: is it power of 2 + 1?
+  * are vertex positions (object-space) valid and not out-of-bounds?
+* Feedback on Success and Error: show in status bar oder as tooltip over button (latter only for errors)
+
+
+### Changed
+
+* GUI Layout
+
+
+### Fixed
+
+* When selecting another object, the stats were not updated.
+
+
 ## [1.1.0] - 2023-04-24
 
 ### Added